hcc seems to require 1024 work item dims.
At least linkobj.cpp test fails silently with a heap overflow when it falls
back to a smaller than 1000 dimension and ends up launching multiple
work groups while wrongly thinking the local size is 1000. This should
be fixed in hcc, but it's easier to bump up the min size in phsa-runtime
as it makes no difference really for a CPU Agent.