Allow launching bare kernels
Such kernels might be finalized from other input than HSAIL and the
finalization API.

The kernel function name must be prefixed with 'phsa_kernel.'

Added one argument to the kernel function, a direct pointer to the
argument buffer, to avoid pointer chasing the phsa context struct
to find it in the kernel function.

Also other smaller fixes and cleanups.